+++ CHANGELOG.md
@@ -25,7 +25,9 @@
 * Update received piece table columns. Refs UIIN-1632.
 * Restore `mod-inventory` endpoints for non-search requests. UIIN-1634.
 * Delete Inventory and MARC holdings records when record Source = MARC. Refs UIIN-1559.
+* Update react-final-form to 6.5.6 to fix breaking change.
 * Increment `stripes` to `v7`, `react` to `v17`. Refs UIIN-1564.
+* Edit MARC Holdings via quickMARC. Refs UIIN-1636.
 * Instance record: Update Instance record Actions menu. Refs UIIN-1625.
 
 ## [7.1.4](https://github.com/folio-org/ui-inventory/tree/v7.1.4) (2021-08-05)